Separate global vars and funcs
