+++ controllers/imagebuilder_test.go
@@ -2,9 +2,6 @@ package controllers
 
 import (
 	"context"
-	"math/rand"
-	"path/filepath"
-	"testing"
 	"time"
 
 	. "github.com/onsi/ginkgo"
@@ -12,68 +9,19 @@ import (
 
 	core "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/client-go/rest"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/envtest"
-	logf "sigs.k8s.io/controller-runtime/pkg/log"
-	"sigs.k8s.io/controller-runtime/pkg/log/zap"
+
 	// +kubebuilder:scaffold:imports
 
 	primehubv1alpha1 "primehub-controller/api/v1alpha1"
 )
 
-// These tests use Ginkgo (BDD-style Go testing framework). Refer to
-// http://onsi.github.io/ginkgo/ to learn more about Ginkgo.
-
-var cfg *rest.Config
-var k8sClient client.Client
-var testEnv *envtest.Environment
-
-func TestAPIs(t *testing.T) {
-	RegisterFailHandler(Fail)
-
-	RunSpecsWithDefaultAndCustomReporters(t,
-		"Controller Suite",
-		[]Reporter{envtest.NewlineReporter{}})
-}
-
-var _ = BeforeSuite(func(done Done) {
-	logf.SetLogger(zap.LoggerTo(GinkgoWriter, true))
-
-	By("bootstrapping test environment")
-	testEnv = &envtest.Environment{
-		CRDDirectoryPaths: []string{filepath.Join("..", "config", "crd", "bases")},
-	}
-
-	var err error
-	cfg, err = testEnv.Start()
-	Expect(err).ToNot(HaveOccurred())
-	Expect(cfg).ToNot(BeNil())
-
-	err = primehubv1alpha1.AddToScheme(scheme.Scheme)
-	Expect(err).NotTo(HaveOccurred())
-
-	// +kubebuilder:scaffold:scheme
-
-	k8sClient, err = client.New(cfg, client.Options{Scheme: scheme.Scheme})
-	Expect(err).ToNot(HaveOccurred())
-	Expect(k8sClient).ToNot(BeNil())
-
-	close(done)
-}, 60)
-
-var _ = AfterSuite(func() {
-	By("tearing down the test environment")
-	err := testEnv.Stop()
-	Expect(err).ToNot(HaveOccurred())
-})
-
-func SetupTest(ctx context.Context) *core.Namespace {
-	var stopCh chan struct{}
+var _ = Context("Inside of a new namespace", func() {
+	ctx := context.TODO()
 	ns := &core.Namespace{}
+
+	var stopCh chan struct{}
 
 	BeforeEach(func() {
 		stopCh = make(chan struct{})
@@ -108,13 +56,6 @@ func SetupTest(ctx context.Context) *core.Namespace {
 		Expect(err).NotTo(HaveOccurred(), "failed to delete test namespace")
 	})
 
-	return ns
-}
-
-var _ = Context("Inside of a new namespace", func() {
-	ctx := context.TODO()
-	ns := SetupTest(ctx)
-
 	Describe("when no existing resources exist", func() {
 
 		It("should create a new ImageSpecJob resource after a new ImageSpec was created", func() {
@@ -147,23 +88,3 @@ var _ = Context("Inside of a new namespace", func() {
 		})
 	})
 })
-
-func getResourceFunc(ctx context.Context, key client.ObjectKey, obj runtime.Object) func() error {
-	return func() error {
-		return k8sClient.Get(ctx, key, obj)
-	}
-}
-
-func init() {
-	rand.Seed(time.Now().UnixNano())
-}
-
-var letterRunes = []rune("abcdefghijklmnopqrstuvwxyz1234567890")
-
-func randStringRunes(n int) string {
-	b := make([]rune, n)
-	for i := range b {
-		b[i] = letterRunes[rand.Intn(len(letterRunes))]
-	}
-	return string(b)
-}
